packages/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentfile"
version = "0.1.0"
description = "CLI for building and deploying AI agents as containers"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["ai", "agents", "cli", "docker", "automation"]
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "docker>=7.0",
    "mcp>=1.0",
    "jsonschema>=4.0",
    "psycopg[binary]>=3.0",
    "httpx>=0.27",
]

[project.scripts]
agentfile = "agentfile.cli:cli"
ninetrix = "agentfile.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentfile*"]

[tool.setuptools.package-data]
"agentfile.templates" = ["*.j2", "*.yaml"]
"agentfile.core" = ["*.json"]
